ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.4.0

# Image processing
Pillow==10.4.0
python-magic==0.4.27
piexif==1.1.3
numpy==1.26.4
opencv-python-headless==4.10.0.84

# AI / ML
torch==2.4.0
torchvision==0.19.0
transformers==4.44.0
grad-cam==1.5.5
einops==0.8.0

# Supabase
supabase==2.7.0

# Celery / Redis queue
celery[redis]==5.4.0
redis==5.0.8

# Observability
sentry-sdk[fastapi]==2.14.0

# Dev / test
pytest==8.3.0
pytest-asyncio==0.23.0
httpx==0.27.0
